Organized project.
Learned about @/app/, which is cool.
