+++ app/cabins/page.js
@@ -1,19 +1,7 @@
-import Counter from "../components/Counter";
-
-export default async function Page() {
-  const res = await fetch("https://jsonplaceholder.typicode.com/users");
-  const data = await res.json();
-
+export default function Page() {
   return (
     <div>
       <h1>Cabins page</h1>
-      <ul>
-        {data.map((user) => (
-          <li key={user.id}>{user.name}</li>
-        ))}
-      </ul>
-
-      <Counter users={data} />
     </div>
   );
 }
